+++ src/lib/components/state/ZeroState.tsx
@@ -4,6 +4,7 @@ import { CustomIcon } from "../icon";
 import { useInternalNavigate } from "lib/app-provider";
 import { SaveNewContractModal } from "lib/components/modal/contract";
 import { ADMIN_SPECIAL_SLUG, INSTANTIATED_LIST_NAME } from "lib/data";
+import { SaveAccountButton } from "lib/pages/saved-accounts/components/SaveAccountButton";
 import type { LVPair } from "lib/types";
 import { formatSlugName } from "lib/utils";
 
@@ -86,3 +87,26 @@ export const ZeroState = ({ list, isReadOnly }: ZeroStateProps) => {
     </Flex>
   );
 };
+
+export const AccountZeroState = () => (
+  <Flex
+    alignItems="center"
+    gap={4}
+    color="text.dark"
+    direction="column"
+    my={12}
+    py={8}
+    borderY="1px solid"
+    borderColor="gray.700"
+  >
+    <StateImage imageVariant="empty" />
+    <Flex align="center">You don’t have any saved accounts.</Flex>
+    <Flex align="center" gap={4}>
+      Save an account and entering the account address through
+      <SaveAccountButton />
+    </Flex>
+    <Flex align="center">
+      Saved accounts are stored locally on your device..
+    </Flex>
+  </Flex>
+);