+++ src/lib/stores/account.ts
@@ -78,4 +78,21 @@ export class AccountStore {
   isAccountSaved(address: Addr): boolean {
     return this.savedAccounts[this.userKey]?.includes(address) ?? false;
   }
+
+  removeSavedAccount(address: Addr): void {
+    this.savedAccounts[this.userKey] = this.savedAccounts[this.userKey]?.filter(
+      (each) => each !== address
+    );
+  }
+
+  getSavedAccounts(): AccountLocalInfo[] {
+    const savedAccountsByUserKey = this.savedAccounts[this.userKey] ?? [];
+    return savedAccountsByUserKey
+      .map((address) => ({
+        address,
+        name: this.getAccountLocalInfo(address)?.name,
+        description: this.getAccountLocalInfo(address)?.description,
+      }))
+      .reverse();
+  }
 }